Move DictionaryEqualityComparer -> collections namespace
